mcp-confluence/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.25
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.27.0
llama-cloud>=0.1.0
apscheduler>=3.10.4
python-multipart>=0.0.9
beautifulsoup4>=4.12.0
markdownify>=0.12.0
python-dotenv>=1.0.0
pydantic>=2.9.0